+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.1.0
 
 - Added setting to disable ticket comment logging
+- Fixed project-channel sharding to work as described
 
 ### 1.0.1